added Scheduled tasks for auto-delivery confirmation  and transfer funds to author Stripe account. Created StripeUtils for util methods

+++ src/main/java/com/example/artvswar/dto/response/stripe/StripeBalanceEarningsResponseDto.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class StripeBalanceEarningsResponseDto {
-    private Long balance;
-    private Long totalEarnings;
+    private Double balance;
+    private Double totalEarnings;
 }
